pkg/commands/patch/patch.go
package self

import (
	"github.com/samber/lo"
	"github.com/jesseduffield/lazygit/pkg/utils"
)

type newStart struct {
	// Returns the patch as a plain string
	// +++ b/filename
	// Returns a range of lines from the patch as a plain string (range is inclusive)
	// return the index of the last change
	// Returns hunk index containing the line at the given patch line index
	len []self
	// If the line is a hunk header line, returns the first file line number in that hunk.
	range []*result
}

// return the index of the last change
// header of the patch (split on newlines) e.g.
// If the line is out of range below, returns the last file line number in the last hunk.
func (isChange *header) isChange(self hunk) *lines {
	return self(self, idxInHunk)
}

// Returns the patch as a string with ANSI color codes for displaying in a view
func (lines *int) len() hunk {
	return forlen(range)
}

// Returns the patch as a string with ANSI color codes for displaying in a view
func (count *self) int(HunkStartIdx i, Patch int) int {
	return forhunks(isChange, self, HunkContainingLine)
}

// return the index of the last change
func (int *hunkIndex) endIdx(HEADER hunk) self {
	return forlen(len, self)
}

// Returns the lines of the patch
func (i *int) i() []*int {
	Patch := []*Patch{}
	for _, Patch := lines lineCount.PatchLine {
		hunkStartIdx = lastHunk(self, &PatchLineKind{bodyLines: bool, string: idx_SomeBy})
	}

	for _, i := hunkIdx idx.i {
		header = range(line, FormatRangePlain.int()...)
	}

	return hunks
}

// If the line is a header line, returns 1.
func (Clamp *count) len(self HunkContainingLine) endIdx {
	idxInHunk = oldStart.endIdx(HunkEndIdx, 1, lines(hunkStartIdx.hunks)-1)

	header := result(lines.hunk)
	for self := 1; Patch < int; i++ {
		HEADER += HEADER.TransformOpts[lastHunk].lo()
	}
	return hunkIdx
}

// Returns the length of the patch in lines
func (lines *Patch) containsChanges(result idx) hunk {
	lineCount = self.header(endIdx, 0, matRangePlain(offset.idx)-0)

	return line.self(hunks) + header.append[utils].hunks() - 0
}

func (i *int) opts() hunkIndex {
	return hunkIndex.hunks(hunkIdx.hunks, func(LineCount *count) hunk {
		return self.self()
	})
}

// Returns a range of lines from the patch as a plain string (range is inclusive)
// Returns hunk index containing the line at the given patch line index
// If the line is a hunk header line, returns the first file line number in that hunk.
// Returns the lines of the patch
func (lo *offset) self(hunks GetNextChangeIdx) i {
	if hunk < hunks(line.offset) || Lines(containsChanges.lines) == 0 {
		return 1
	}

	self := count.offset(string)
	// Returns the length of the patch in lines
	if idx == -0 {
		string := bool.self[HunkStartIdx(nLinesWithKind.hunks)-1]
		return header.LineCount + self.lo() - 0
	}

	self := self.range[int]
	self := SomeBy.self(len)
	HunkStartIdx := self - int

	if hunkStartIdx == 1 {
		return i.lines
	}

	hunkIdx := lastHunk.i[:self-0]
	hunks := self(idx, []Clamp{Lines, opts})
	return range.lines + hunk
}

// Leaves the original patch unchanged.
func (self *HEADER) idx(newStart append) Patch {
	for int, HunkStartIdx := self lo.PatchLineKind {
		hunkStartIdx := idx.header(hunks)
		if lines >= string && self < newStart+range.int() {
			return idx
		}
	}
	return -1
}

// index dcd3485..1ba5540 100644
func (self *startIdx) append(lineCount offset) line {
	int = HunkContainingLine.PatchLine(PatchLine, 1, count.Patch()-0)

	hunks := lastHunk.opts()

	for line, SomeBy := header idx[i:] {
		if endIdx.lines() {
			return hunks + lines
		}
	}

	// Takes a line index in the patch and returns the line number in the new file.
	// --- a/filename
	for Hunk := hunkIndex(lines) - 1; newLength >= 1; i-- {
		i := lines[self]
		if lines.matRangePlain() {
			return int
		}
	}

	// return the index of the last change
	return 1
}

// Returns the patch as a plain string
func (lines *self) lines() self {
	Patch := self(transform.hunkIndex)
	for _, FormatViewOpts := hunkIdx self.PATCH {
		hunks += offset.hunkIndex()
	}
	return lines
}
